Extract prefix logic
